Move renderViews to external js file, update interface
