+++ src/style.css
@@ -3,6 +3,13 @@ body {
   padding: 0;
   background-color: cadetblue;
 }
+
+#app {
+  height: 100vh;
+  display: flex;
+  flex-direction: column;
+}
+
 .main-forecast {
   display: flex;
   flex-flow: row;
@@ -10,12 +17,28 @@ body {
   overflow: auto;
 }
 .forecast-item {
-  margin: 0 1rem;
+  padding: 0 2rem;
+  display: flex;
+  flex-direction: column;
+  justify-content: flex-end;
+  align-items: center;
+
 }
 
 #current-weather {
   line-height: 1.3rem;
   text-align: center;
+  display: flex;
+  justify-content: center;
+  flex-grow: 1;
+  margin-bottom: 275px;
+  align-items: center;
+  border-bottom: 3px dotted #d6c902;
+  border-top: 2px dotted #d6c902;
+}
+
+.current-weather-nodes {
+  margin: 0 1.3rem;
 }
 
 .location {
@@ -25,8 +48,12 @@ body {
 }
 
 #current-forecast {
+  height: 275px;
   display: flex;
   overflow: auto;
+  position: fixed;
+  bottom: 0;
+  width: 100%;
   white-space: nowrap;
 }
